scripts/color_generator/requirements.txt
wheel
fire
psycopg2-binary
